// src/app/generated-apis/user/model/tattooStyle.ts
/**
 * user-generated-contract
 * user-generated-contract
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TattooStyle = 'OLD_SCHOOL' | 'TRIBAL' | 'ASIAN_ORIENTAL' | 'SCRIPT' | 'WATERCOLOUR' | 'PORTRAITS' | 'NEW_SCHOOL' | 'BLACK_AND_GREY' | 'BIOMECHANICAL' | 'REALISTIC' | 'DOT_WORK';

export const TattooStyle = {
    OldSchool: 'OLD_SCHOOL' as TattooStyle,
    Tribal: 'TRIBAL' as TattooStyle,
    AsianOriental: 'ASIAN_ORIENTAL' as TattooStyle,
    Script: 'SCRIPT' as TattooStyle,
    Watercolour: 'WATERCOLOUR' as TattooStyle,
    Portraits: 'PORTRAITS' as TattooStyle,
    NewSchool: 'NEW_SCHOOL' as TattooStyle,
    BlackAndGrey: 'BLACK_AND_GREY' as TattooStyle,
    Biomechanical: 'BIOMECHANICAL' as TattooStyle,
    Realistic: 'REALISTIC' as TattooStyle,
    DotWork: 'DOT_WORK' as TattooStyle
};